requirements.txt
openai>=1.40.0
anthropic>=0.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
fake-useragent>=1.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
loguru>=0.7.0